+++ CHANGELOG.md
@@ -8,7 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 **Added**
 
-* Added new regions for cloud service: LIN, MTZ, VCP, BRS, UKB
+
+* Added new regions for cloud service: LIN, MTZ, VCP, BRS, UKB.
+* Added Instance Principal with Delegation Token feature.
 * Added NoSQLClient.queryIterable() API to iterate over query results using
 for-await-of loop.